Solve phpunit. Fix database field format

+++ htdocs/install/mysql/migration/3.7.0-3.8.0.sql
@@ -47,9 +47,9 @@ ALTER TABLE llx_product_fournisseur_price ADD fk_price_expression integer DEFAUL
 insert into llx_c_tva(rowid,fk_pays,taux,recuperableonly,note,active) values ( 2131, 213, '5', '0', 'VAT 5%', 1);
 
 -- Add situation invoices
-ALTER TABLE llx_facture ADD COLUMN situation_cycle_ref integer;
-ALTER TABLE llx_facture ADD COLUMN situation_counter integer;
-ALTER TABLE llx_facture ADD COLUMN situation_final integer;
+ALTER TABLE llx_facture ADD COLUMN situation_cycle_ref smallint;
+ALTER TABLE llx_facture ADD COLUMN situation_counter smallint;
+ALTER TABLE llx_facture ADD COLUMN situation_final smallint;
 ALTER TABLE llx_facturedet ADD COLUMN situation_percent real;
 ALTER TABLE llx_facturedet ADD COLUMN fk_prev_id integer;
